feat:  def transaction_result(self, coins, cost): This is used to determine whether enough money was given to complete the purchase or not and returns and error message if there is not.
fix: When being asked to select sandwich size, originally had all capital words but then made it invalid so had to make them lowercase.

feat: def process_coins(self): This feature sends the question of how many of each type of money and then has an initialized value to multiply with the amount.

feat:     if user_input == "off": >> break  : This feature allows the user to enter the word "off" and it will shut down the compiler and immediately end the process.

feat: def check_resources(self, ingredients): Allows the system to check the inventory for the required materials needed for each sandwich and will return a statement if there is not enough.